Mark tests as ported to the TCK
- Delete inappropriately named test file
- Move tests to other files with more alike tests
